Fix Bugs and Refactor
