+++ PvPController/Network/PacketArgs/PlayerSlotArgs.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PvPController.Network {
+namespace PvPController.Network.PacketArgs {
     public class PlayerSlotArgs : EventArgs {
         public PvPPlayer player;
         public int slotid;
