+++ PvPController/Network/PacketArgs/ProjectileDestroyArgs.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PvPController.Network {
+namespace PvPController.Network.PacketArgs {
     public class ProjectileDestroyArgs : EventArgs {
         public int projectileIndex { get; set; }
         public int owner { get; set; }
