+++ PvPController/PvPController.csproj
@@ -67,22 +67,22 @@
     <Compile Include="Database.cs" />
     <Compile Include="Interface.cs" />
     <Compile Include="Network\DataHandler.cs" />
-    <Compile Include="Network\PlayerDeathArgs.cs" />
-    <Compile Include="Network\PlayerHurtArgs.cs" />
-    <Compile Include="Network\PlayerSlotArgs.cs" />
-    <Compile Include="Network\PlayerUpdateArgs.cs" />
-    <Compile Include="Network\ProjectileDestroyArgs.cs" />
-    <Compile Include="Network\ProjectileNewArgs.cs" />
+    <Compile Include="Network\PacketArgs\PlayerDeathArgs.cs" />
+    <Compile Include="Network\PacketArgs\PlayerHurtArgs.cs" />
+    <Compile Include="Network\PacketArgs\PlayerSlotArgs.cs" />
+    <Compile Include="Network\PacketArgs\PlayerUpdateArgs.cs" />
+    <Compile Include="Network\PacketArgs\ProjectileDestroyArgs.cs" />
+    <Compile Include="Network\PacketArgs\ProjectileNewArgs.cs" />
     <Compile Include="Network\PvPHandler.cs" />
-    <Compile Include="Network\TogglePvPArgs.cs" />
-    <Compile Include="Utilities\MinionUtils.cs" />
-    <Compile Include="Utilities\MiscData.cs" />
+    <Compile Include="Network\PacketArgs\TogglePvPArgs.cs" />
+    <Compile Include="Utilities\PresetData.cs" />
     <Compile Include="PluginCommands.cs" />
     <Compile Include="PvPController.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utilities\ProjectileUtils.cs" />
     <Compile Include="Utilities\PvPUtils.cs" />
     <Compile Include="Variables\BuffDuration.cs" />
+    <Compile Include="Variables\MinionProjectile.cs" />
     <Compile Include="Variables\PvPItem.cs" />
     <Compile Include="Variables\PvPPlayer.cs" />
     <Compile Include="Variables\PvPProjectile.cs" />
